LPS-197820 move constants to api
