+++ modules/apps/saved-content/saved-content-api/src/main/java/com/liferay/saved/content/constants/MySavedContentPortletKeys.java
@@ -3,10 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.saved.content.web.internal.constants;
+package com.liferay.saved.content.constants;
 
 /**
- /**
  * @author Alicia García
  */
 public class MySavedContentPortletKeys {
